requirements.txt
avalanche-lib>=0.6.0
qpsolvers>=4.3.2
pandas
scikit-learn
torch